test(store): split the store layer's self-test by module under test
.github/store/tests/test_store.py had reached 2222 lines across 22
unittest classes covering seven modules. Each module's classes get
their own file — preferences, config, guards, replay, extraction,
similarity — and the sys.path bootstrap and record factories they all
share move to store_support.py beside them.

One file was the entry point, so the runner changes with it: the
preferences-guard workflow, the store README and the template's own
copy of the suite discover the directory instead of naming a file.
Same 181 tests, no assertion changed.

ADVANCES pandoscope/agentic-engineering-template#239.
