+++ tests/test_decision_memory_subtemplate.py
@@ -34,7 +34,13 @@ STORE_FILES = frozenset(
         ".github/store/render_preferences.py",
         ".github/store/replay.py",
         ".github/store/similarity.py",
-        ".github/store/tests/test_store.py",
+        ".github/store/tests/store_support.py",
+        ".github/store/tests/test_config.py",
+        ".github/store/tests/test_extraction.py",
+        ".github/store/tests/test_guards.py",
+        ".github/store/tests/test_preferences.py",
+        ".github/store/tests/test_replay.py",
+        ".github/store/tests/test_similarity.py",
         ".github/workflows/preferences-budget.yml",
         ".github/workflows/preferences-guard.yml",
         ".github/workflows/guards.yml",
